Fixes to get base quatcam functionality working again including: working solely with dcs, no more cfdnavigate, and matching up datatypes. This is a testing level commit, there's currently a lot of unused code in each class.

+++ VE_Xplorer/XplorerHandlers/cfdQuatCam.h
@@ -73,12 +73,12 @@ namespace VE_Xplorer
    {
       public:
          //Constructors
-         cfdQuatCam( gmtl::Matrix44f&, double* );
+         cfdQuatCam( gmtl::Matrix44f&, float* );
    
          //Destructor
          ~cfdQuatCam(){;}
 
-         void SetCamPos( double*, VE_SceneGraph::DCS* );
+         void SetCamPos( float*, VE_SceneGraph::DCS* );
 
          void MoveCam( float );
